+++ packages/open-next/src/adapters/server-adapter.ts
@@ -12,9 +12,6 @@ setBuildIdEnv();
 setNextjsServerWorkingDirectory();
 
 // Because next is messing with fetch, we have to make sure that we use an untouched version of fetch
-declare global {
-  var internalFetch: typeof fetch;
-}
 globalThis.internalFetch = fetch;
 
 /////////////
