+++ packages/open-next/src/core/createMainHandler.ts
@@ -1,5 +1,3 @@
-import type { AsyncLocalStorage } from "node:async_hooks";
-
 import type { OpenNextConfig } from "types/open-next";
 import type { IncrementalCache, Queue } from "types/overrides";
 import type { DetachedPromiseRunner } from "utils/promise";
@@ -14,19 +12,6 @@ import {
   resolveTagCache,
   resolveWrapper,
 } from "./resolve";
-
-declare global {
-  var queue: Queue;
-  var incrementalCache: IncrementalCache;
-  var fnName: string | undefined;
-  var serverId: string;
-  var __als: AsyncLocalStorage<{
-    requestId: string;
-    pendingPromiseRunner: DetachedPromiseRunner;
-    isISRRevalidation?: boolean;
-    mergeHeadersPriority?: "middleware" | "handler";
-  }>;
-}
 
 export async function createMainHandler() {
   // @ts-expect-error `./open-next.config.mjs` exists only in the build output
